# packages/contracts/openapi/current.yaml
openapi: 3.1.0
info:
  title: Affiliate GLOBAL Current API
  version: 1.0.0-go-rewrite-freeze
  description: Behavioral contract frozen from the running Nest oracle on 2026-07-22.
servers:
  - url: /
paths:
...
  /auth/mock-login:
    post:
      operationId: mockLogin
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MockLoginInput" } } } }
      responses:
        "201": { description: Session issued, content: { application/json: { schema: { $ref: "#/components/schemas/LoginResult" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/me:
    get:
      operationId: getCurrentAuth
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Session-derived identity and roles, content: { application/json: { schema: { $ref: "#/components/schemas/AuthContext" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      operationId: logout
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Session revoked, content: { application/json: { schema: { $ref: "#/components/schemas/Ok" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
...
  responses:
    BadRequest: { description: Invalid input, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
    Unauthorized: { description: Missing or invalid session, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
...
  schemas:
    UUID: { type: string, format: uuid }
    ISODate: { type: string, format: date-time }
...
    Ok:
      type: object
      additionalProperties: false
      required: [ok]
      properties: { ok: { type: boolean, const: true } }
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, status, correlationId, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            status: { type: integer }
            correlationId: { type: string, format: uuid }
            retryable: { type: boolean }
...
    AuthUser:
      type: object
      additionalProperties: false
      required: [id, email, displayName]
      properties: { id: { $ref: "#/components/schemas/UUID" }, email: { type: string }, displayName: { type: string } }
    AuthRole:
      type: object
      additionalProperties: false
      required: [countryId, role]
      properties: { countryId: { type: [string, "null"], format: uuid }, role: { type: string } }
    AuthContext:
      type: object
      additionalProperties: false
      required: [user, roles]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
        roles: { type: array, items: { $ref: "#/components/schemas/AuthRole" } }
    MockLoginInput:
      type: object
      required: [email]
      properties: { email: { type: string, format: email }, displayName: { type: string } }
    LoginResult:
      type: object
      additionalProperties: false
      required: [token, expiresAt, user]
      properties:
        token: { type: string }
        expiresAt: { $ref: "#/components/schemas/ISODate" }
        user: { $ref: "#/components/schemas/AuthUser" }
...